Move App to src/components
