--- src/App.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function App() {
-  return <div className="App">Ephemeral</div>;
-}
-
-export default App;
