+++ src/components/App.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function App () {
+  return <div className='App'>Ephemeral</div>
+}
+
+export default App
